+++ angel_orm/lib/angel_orm.dart
@@ -1,4 +1,3 @@
 export 'src/annotations.dart';
-export 'src/migration.dart';
-export 'src/relations.dart';
 export 'src/query.dart';
+export 'src/relations.dart';